+++ Scintillating.ProxyProtocol.Parser.Tests/ProtocolV1IncompleteParserTests.cs
@@ -139,7 +139,7 @@ public class ProtocolV1IncompleteParserTests
                         proxyProtocolHeader!.Command
                             .Should().Be(ProxyCommand.Proxy, "there is a single command in protocol V1");
                         proxyProtocolHeader!.Length
-                            .Should().Be((ushort)expectedLength, "entire header has to be consumed");
+                            .Should().Be(expectedLength, "entire header has to be consumed");
 
                         proxyProtocolHeader!.AddressFamily
                             .Should().Be(addressFamily, "it should match the sample");
